Clean classes and packages with inheritance
