+++ src/main/java/net/jukoz/me/entity/projectile/pebble/PebbleEntity.java
@@ -2,6 +2,7 @@ package net.jukoz.me.entity.projectile.pebble;
 
 import net.jukoz.me.entity.ModEntities;
 import net.jukoz.me.entity.hobbits.shire.ShireHobbitEntity;
+import net.jukoz.me.entity.projectile.AbstractProjectleEntity;
 import net.jukoz.me.item.ModResourceItems;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
@@ -16,7 +17,7 @@ import net.minecraft.util.hit.EntityHitResult;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.world.World;
 
-public class PebbleEntity extends ThrownItemEntity {
+public class PebbleEntity extends AbstractProjectleEntity {
     private Item item = ModResourceItems.PEBBLE;
     private float damage;
 
@@ -35,33 +36,5 @@ public class PebbleEntity extends ThrownItemEntity {
 
     protected Item getDefaultItem() {
         return item;
-    }
-
-    private ParticleEffect getParticleParameters() {
-        return new ItemStackParticleEffect(ParticleTypes.ITEM, new ItemStack(item));
-    }
-
-    public void handleStatus(byte status) {
-        if (status == 3) {
-            ParticleEffect particleEffect = this.getParticleParameters();
-            for(int i = 0; i < 8; ++i) {
-                getWorld().addParticle(particleEffect, this.getX(), this.getY(), this.getZ(), 0.0, 0.0, 0.0);
-            }
-        }
-    }
-
-    protected void onEntityHit(EntityHitResult entityHitResult) {
-        super.onEntityHit(entityHitResult);
-        Entity entity = entityHitResult.getEntity();
-        if(this.getOwner() instanceof ShireHobbitEntity && entity instanceof ShireHobbitEntity) return;
-        entity.damage(this.getDamageSources().thrown(this, this.getOwner()), this.damage);
-    }
-
-    protected void onCollision(HitResult hitResult) {
-        super.onCollision(hitResult);
-        if (!getWorld().isClient) {
-            getWorld().sendEntityStatus(this, (byte)3);
-            this.discard();
-        }
     }
 }
