+++ src/main/java/net/jukoz/me/item/items/LitPineconeItem.java
@@ -1,19 +1,15 @@
 package net.jukoz.me.item.items;
 
-import net.jukoz.me.entity.projectile.lit_pinecone.LitPineconeEntity;
+import net.jukoz.me.entity.projectile.pinecone.LitPineconeEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.stat.Stats;
-import net.minecraft.util.Hand;
-import net.minecraft.util.TypedActionResult;
-import net.minecraft.util.UseAction;
 import net.minecraft.world.World;
 
-public class LitPineconeItem extends Item{
+public class LitPineconeItem extends PineconeItem{
     public static final float DAMAGE = 2f;
     private static final float BASE_STRENGTH = 0.6f;
     private static final float CHARGE_STRENGTH = 0.9f;
@@ -22,21 +18,6 @@ public class LitPineconeItem extends Item{
         super(settings);
     }
 
-    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-        ItemStack itemStack = user.getStackInHand(hand);
-        user.setCurrentHand(hand);
-        return TypedActionResult.success(itemStack, world.isClient());
-    }
-
-    @Override
-    public UseAction getUseAction(ItemStack stack) {
-        return UseAction.BOW;
-    }
-
-    @Override
-    public int getMaxUseTime(ItemStack stack) {
-        return 96000;
-    }
 
     @Override
     public void onStoppedUsing(ItemStack stack, World world, LivingEntity user, int remainingUseTicks) {
