+++ src/main/java/net/jukoz/me/item/items/PineconeItem.java
@@ -13,29 +13,13 @@ import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.UseAction;
 import net.minecraft.world.World;
 
-public class PineconeItem extends Item{
+public class PineconeItem extends PebbleItem{
     public static final float DAMAGE = 2f;
     private static final float BASE_STRENGTH = 0.6f;
     private static final float CHARGE_STRENGTH = 0.9f;
     private static final int STRENGTH_CHARGE_TIME = 20; // 1s charge for full strength
     public PineconeItem(Settings settings) {
         super(settings);
-    }
-
-    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-        ItemStack itemStack = user.getStackInHand(hand);
-        user.setCurrentHand(hand);
-        return TypedActionResult.success(itemStack, world.isClient());
-    }
-
-    @Override
-    public UseAction getUseAction(ItemStack stack) {
-        return UseAction.BOW;
-    }
-
-    @Override
-    public int getMaxUseTime(ItemStack stack) {
-        return 72000;
     }
 
     @Override
